Move common methods to helpers file
