+++ src/libvirt-dbus.js
@@ -100,65 +100,12 @@ import {
     updateNetworkIface,
     updateVCPUSettings,
 } from './libvirt-xml-update.js';
-
-const clientLibvirt = {};
-/* Default timeout for libvirt-dbus method calls */
-const timeout = 30000;
-
-const Enum = {
-    VIR_DOMAIN_AFFECT_CURRENT: 0,
-    VIR_DOMAIN_AFFECT_LIVE: 1,
-    VIR_DOMAIN_AFFECT_CONFIG: 2,
-    VIR_DOMAIN_INTERFACE_ADDRESSES_SRC_LEASE: 0,
-    VIR_DOMAIN_INTERFACE_ADDRESSES_SRC_AGENT: 1,
-    VIR_DOMAIN_INTERFACE_ADDRESSES_SRC_ARP: 2,
-    VIR_DOMAIN_UNDEFINE_MANAGED_SAVE: 1,
-    VIR_DOMAIN_UNDEFINE_SNAPSHOTS_METADATA: 2,
-    VIR_DOMAIN_UNDEFINE_NVRAM: 4,
-    VIR_DOMAIN_SNAPSHOT_LIST_INTERNAL : 256,
-    VIR_DOMAIN_STATS_BALLOON: 4,
-    VIR_DOMAIN_SHUTOFF: 5,
-    VIR_DOMAIN_STATS_VCPU: 8,
-    VIR_DOMAIN_STATS_BLOCK: 32,
-    VIR_DOMAIN_STATS_STATE: 1,
-    VIR_DOMAIN_XML_SECURE: 1,
-    VIR_DOMAIN_XML_INACTIVE: 2,
-    VIR_CONNECT_LIST_DOMAINS_PERSISTENT: 4,
-    VIR_CONNECT_LIST_DOMAINS_TRANSIENT: 8,
-    VIR_CONNECT_LIST_INTERFACES_INACTIVE: 1,
-    VIR_CONNECT_LIST_INTERFACES_ACTIVE: 2,
-    VIR_CONNECT_LIST_NETWORKS_ACTIVE: 2,
-    VIR_CONNECT_LIST_STORAGE_POOLS_ACTIVE: 2,
-    VIR_CONNECT_LIST_STORAGE_POOLS_DIR: 64,
-    VIR_STORAGE_POOL_CREATE_NORMAL: 0,
-    VIR_STORAGE_POOL_DELETE_NORMAL: 0,
-    // Storage Pools Event Lifecycle Type
-    VIR_STORAGE_POOL_EVENT_DEFINED: 0,
-    VIR_STORAGE_POOL_EVENT_UNDEFINED: 1,
-    VIR_STORAGE_POOL_EVENT_STARTED: 2,
-    VIR_STORAGE_POOL_EVENT_STOPPED: 3,
-    VIR_STORAGE_POOL_EVENT_CREATED: 4,
-    VIR_STORAGE_POOL_EVENT_DELETED: 5,
-    VIR_STORAGE_POOL_EVENT_LAST: 6,
-    VIR_STORAGE_VOL_DELETE_NORMAL: 0,
-    VIR_STORAGE_VOL_DELETE_WITH_SNAPSHOTS: 2,
-    // Networks Event Lifecycle Type
-    VIR_NETWORK_EVENT_DEFINED: 0,
-    VIR_NETWORK_EVENT_UNDEFINED: 1,
-    VIR_NETWORK_EVENT_STARTED: 2,
-    VIR_NETWORK_EVENT_STOPPED: 3,
-    VIR_NETWORK_EVENT_LAST: 4,
-    // Keycodes
-    VIR_KEYCODE_SET_LINUX: 0,
-    // Migrate
-    // https://libvirt.org/html/libvirt-libvirt-domain.html#virDomainMigrateFlags
-    VIR_MIGRATE_LIVE: 1,
-    VIR_MIGRATE_PEER2PEER: 2,
-    VIR_MIGRATE_PERSIST_DEST: 8,
-    VIR_MIGRATE_UNDEFINE_SOURCE: 16,
-    VIR_MIGRATE_NON_SHARED_DISK: 64,
-    VIR_MIGRATE_OFFLINE: 1024,
-};
+import {
+    call,
+    dbusClient,
+    Enum,
+    timeout,
+} from './libvirtApi/helpers.js';
 
 const LIBVIRT_DBUS_PROVIDER = {
     name: 'LibvirtDBus',
@@ -499,7 +446,7 @@ export function getAllVms({ connectionName }) {
 
 export function getApiData({ connectionName, libvirtServiceName }) {
     if (connectionName) {
-        dbus_client(connectionName);
+        dbusClient(connectionName);
         startEventMonitor({ connectionName, libvirtServiceName });
         return Promise.allSettled([
             getAllVms({ connectionName }),
@@ -1053,7 +1000,7 @@ function startEventMonitor({ connectionName }) {
 
 function startEventMonitorDomains(connectionName) {
     /* Subscribe to Domain Lifecycle signals on Connect Interface */
-    dbus_client(connectionName).subscribe(
+    dbusClient(connectionName).subscribe(
         { interface: 'org.libvirt.Connect', member: 'DomainEvent' },
         (path, iface, signal, args) => {
             const domainEvent = {
@@ -1113,7 +1060,7 @@ function startEventMonitorDomains(connectionName) {
     );
 
     /* Subscribe to signals on Domain Interface */
-    dbus_client(connectionName).subscribe(
+    dbusClient(connectionName).subscribe(
         { interface: 'org.libvirt.Domain' },
         (path, iface, signal, args) => {
             logDebug(`signal on ${path}: ${iface}.${signal}(${JSON.stringify(args)})`);
@@ -1183,7 +1130,7 @@ function networkUpdateOrDelete(connectionName, netPath) {
 }
 
 function startEventMonitorNetworks(connectionName) {
-    dbus_client(connectionName).subscribe(
+    dbusClient(connectionName).subscribe(
         { interface: 'org.libvirt.Connect', member: 'NetworkEvent' },
         (path, iface, signal, args) => {
             const objPath = args[0];
@@ -1207,7 +1154,7 @@ function startEventMonitorNetworks(connectionName) {
     );
 
     /* Subscribe to signals on Network Interface */
-    dbus_client(connectionName).subscribe(
+    dbusClient(connectionName).subscribe(
         { interface: 'org.libvirt.Network' },
         (path, iface, signal, args) => {
             switch (signal) {
@@ -1222,7 +1169,7 @@ function startEventMonitorNetworks(connectionName) {
 }
 
 function startEventMonitorStoragePools(connectionName) {
-    dbus_client(connectionName).subscribe(
+    dbusClient(connectionName).subscribe(
         { interface: 'org.libvirt.Connect', member: 'StoragePoolEvent' },
         (path, iface, signal, args) => {
             const objPath = args[0];
@@ -1250,7 +1197,7 @@ function startEventMonitorStoragePools(connectionName) {
     );
 
     /* Subscribe to signals on StoragePool Interface */
-    dbus_client(connectionName).subscribe(
+    dbusClient(connectionName).subscribe(
         { interface: 'org.libvirt.StoragePool' },
         (path, iface, signal, args) => {
             switch (signal) {
@@ -1264,30 +1211,9 @@ function startEventMonitorStoragePools(connectionName) {
         });
 }
 
-/**
- * Get Libvirt D-Bus client
- */
-function dbus_client(connectionName) {
-    if (!(connectionName in clientLibvirt) || clientLibvirt[connectionName] === null) {
-        const opts = { bus: connectionName };
-        if (connectionName === 'system')
-            opts.superuser = 'try';
-        clientLibvirt[connectionName] = cockpit.dbus("org.libvirt", opts);
-    }
-
-    return clientLibvirt[connectionName];
-}
-
 export function getLibvirtVersion({ connectionName }) {
     return call(connectionName, "/org/libvirt/QEMU", "org.freedesktop.DBus.Properties", "Get", ["org.libvirt.Connect", "LibVersion"], { timeout, type: 'ss' })
             .then(version => store.dispatch(updateLibvirtVersion({ libvirtVersion: version[0].v })));
-}
-
-/**
- * Call a Libvirt method
- */
-function call(connectionName, objectPath, iface, method, args, opts) {
-    return dbus_client(connectionName).call(objectPath, iface, method, args, opts);
 }
 
 function attachDevice({ connectionName, vmId, permanent, hotplug, xmlDesc }) {
